GH Actions/linting: show deprecations when linting
While rare, there are some deprecations which PHP can show when a file is being linted.
By default these are ignored by PHP-Parallel-Lint.

However, there is an option to show them, so let's turn that option on.

To make the option effective, we also need to ensure that PHP is run in a way that all errors will be shown, which is not the case by default, so this also sets some `ini-values` for PHP to a more optimal setting for CI.

Lastly, in the command, let's also make the package directory iteration a little faster by preventing parallel lint from having to iterate through the `.git` directory when there are no PHP files to be found there anyway.